Update Horde demos and HordeHelper for Horde 1.0.0-Beta5 + PTB 3.0.12
As if by magic, HordeHelper('GetRenderTarget') and HordeHelper('EndOpenGL')
are no longer needed. The only needed call is HordeHelper('Initialize').

All the buffer switching is done automagically in all modes,
as of the Screen() mex file for PTB 3.0.12 and of
Horde3D 1.0.0-Beta 5 Git master.

Cleanup HorderHelper, cleanup and simplify all the demos.

-> Simplify your life.

-> Tested for standard non-imaging mode operation, imaging pipeline,
   and rendering to fast offscreen windows + display with imaging
   pipeline.