Remove mex files/libs for all dead and unsupported platforms.
Windows 32-Bit is dead. Linux 32-Bit is not yet dead, but the
files here were totally outdated, and i am not in the mood to
go through all the moves for 32-Bit Linux support. Given how
easy it is to recreate 32-Bit Mex files for 32-Bit Octave on
Linux, this is left as an exercise to the reader.

-> Horde3DCore mex files up to date for 64-Bit Matlab on Windows,
   64-Bit Matlab/Octave on OSX, and 64-Bit Matlab/Octave on Linux.

-> Demos in a good shape, even the VR HMD demos.