makeHorde/OSX: Update for recent PTB and macOS. Some Apple Silicon prep.
Note that for some fracking reason we now need to use install_name_tool to change
the absolute path of the Horde3D framework libraries to the mex files @loader_path.

Why? No clue, just a new thing we need, now that we are building under macOS 13,
or with XCode 15.2. Another 3 hours of my life I'll never get back...